+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ CHANGELOG
 0.2
 ---
 
+* Renames SQLStore to SourceLoader
+* Improvements to SQLStore
 * Improved documentation
 * Renamed PostgresCopy to PostgresCopyFrom
 * SQLUpload and PostgresCopy have now the same API